+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Bug fixes:
 * Fix `Enumerator::Lazy#{chunk_while, slice_before, slice_after, slice_when}` to return instances of `Enumerator::Lazy` (#2273).
 * Fix `Truffle::Interop.source_location` to return unavailable source sections for modules instead of null (#2257).
 * Fix usage of `Thread.handle_interrupt` in `MonitorMixin#mon_synchronize`.
+* Fixed `TruffleRuby.synchronized` to handle guest safepoints (#2277).
 
 Compatibility: